+++ src/mesh/forest/logical_location.hpp
@@ -227,49 +227,6 @@ inline bool operator!=(const LogicalLocation &lhs, const LogicalLocation &rhs) {
   return !(lhs == rhs);
 }
 
-struct block_ownership_t {
- public:
-  KOKKOS_FORCEINLINE_FUNCTION
-  const bool &operator()(int ox1, int ox2, int ox3) const {
-    return ownership[ox1 + 1][ox2 + 1][ox3 + 1];
-  }
-  KOKKOS_FORCEINLINE_FUNCTION
-  bool &operator()(int ox1, int ox2, int ox3) {
-    return ownership[ox1 + 1][ox2 + 1][ox3 + 1];
-  }
-
-  KOKKOS_FORCEINLINE_FUNCTION
-  block_ownership_t() : block_ownership_t(false) {}
-
-  KOKKOS_FORCEINLINE_FUNCTION
-  explicit block_ownership_t(bool value) : initialized(false) {
-    for (int i = 0; i < 3; ++i) {
-      for (int j = 0; j < 3; ++j) {
-        for (int k = 0; k < 3; ++k) {
-          ownership[i][j][k] = value;
-        }
-      }
-    }
-  }
-
-  bool initialized;
-
-  bool operator==(const block_ownership_t &rhs) const {
-    bool same = initialized == rhs.initialized;
-    for (int i = 0; i < 3; ++i) {
-      for (int j = 0; j < 3; ++j) {
-        for (int k = 0; k < 3; ++k) {
-          same = same && (ownership[i][j][k] == rhs.ownership[i][j][k]);
-        }
-      }
-    }
-    return same;
-  }
-
- private:
-  bool ownership[3][3][3];
-};
-
 struct NeighborLocation {
   NeighborLocation(const LogicalLocation &g, const LogicalLocation &o)
       : global_loc(g), origin_loc(o) {}
@@ -277,27 +234,6 @@ struct NeighborLocation {
   LogicalLocation
       origin_loc; // Logical location of neighboring block in index space of origin block
 };
-
-block_ownership_t
-DetermineOwnership(const LogicalLocation &main_block,
-                   const std::unordered_set<LogicalLocation> &allowed_neighbors,
-                   const RootGridInfo &rg_info = RootGridInfo(),
-                   const std::unordered_set<LogicalLocation> &newly_refined = {});
-
-block_ownership_t
-DetermineOwnershipForest(const LogicalLocation &main_block,
-                         const std::vector<NeighborLocation> &allowed_neighbors,
-                         const std::unordered_set<LogicalLocation> &newly_refined = {});
-
-// Given a topological element, ownership array of the sending block, and offset indices
-// defining the location of an index region within the block (i.e. the ghost zones passed
-// across the x-face or the ghost zones passed across the z-edge), return the index range
-// masking array required for masking out unowned regions of the index space. ox? defines
-// buffer location on the owner block
-block_ownership_t
-GetIndexRangeMaskFromOwnership(TopologicalElement el,
-                               const block_ownership_t &sender_ownership, int ox1,
-                               int ox2, int ox3);
 
 } // namespace parthenon
